SimpleRAG/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
lxml==5.3.0
psycopg2-binary==2.9.10
python-docx==1.1.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
